feat: move chat button composables to commons KMP module
- Move JoinChatButton and LeaveChatButton (ephemChat + nip28) to commons
- Create INav interface in commons, Android INav extends it
- Add ButtonPadding and HalfHalfHorzModifier to commons theme
- Add join/leave string resources to commons
- Convert stringRes(R.string.*) to stringResource(Res.string.*)
- Update Android callers to import from commons

These composables now only depend on commons types (IAccountViewModel,
EphemeralChatChannel, PublicChatChannel) and can be used on iOS/desktop.
